chore: Updated all spells type to float

+++ include/spellslib/champion.hpp
@@ -7,23 +7,23 @@
 
 class ChampionSpells {
 private:
-    std::map<std::string, std::function<double()>> attributes;
+    std::map<std::string, std::function<float()>> attributes;
     
 protected:
     auto create_map() const {
-        return std::map<std::string, std::function<double()>>();
+        return std::map<std::string, std::function<float()>>();
     }
     
-    double find_property_from_map(const std::map<std::string, std::function<double()>> &map, const std::string& property_name) const {
+    float find_property_from_map(const std::map<std::string, std::function<float()>> &map, const std::string& property_name) const {
         auto it = map.find(property_name);
         if (it != map.end())
             return it->second();
-        return 0;
+        return 0.0f;
     }
 
 public:
     ChampionSpells(
-        std::map<std::string, std::function<double()>> attributes = std::map<std::string, std::function<double()>>()
+        std::map<std::string, std::function<float()>> attributes = std::map<std::string, std::function<float()>>()
     ) : attributes(attributes) {}
     virtual ~ChampionSpells() = default;
     
@@ -32,20 +32,20 @@ public:
     virtual Spell get_e() const = 0;
     virtual Spell get_r() const = 0;
     
-    void set_attribute(std::string field, std::function<double()> value) {
+    void set_attribute(std::string field, std::function<float()> value) {
         this->attributes[field] = value;
     }
     
-    double find_property(const std::string& property_name) const {
+    float find_property(const std::string& property_name) const {
         return find_property_from_map(this->attributes, property_name);
     }
     
     void reset_attribute(const std::string& field) {
-        this->attributes[field] = []() -> double { return 0; };
+        this->attributes[field] = []() -> float { return 0.0f; };
     }
     
     void reset_attributes() {
         for(auto& pair : this->attributes)
-            pair.second = []() -> double { return 0; };
+            pair.second = []() -> float { return 0.0f; };
     }
 };
